Data.Cyp

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

newtype Parser s a

The Parser type.

Constructors

Parser ([s] → Maybe (a, [s]))

($>) ∷ Parser s a → [s] → Maybe (a, [s])

Apply a Parser to a Stream.

(|>) ∷ Parser s a → (a → Parser s b) → Parser s b

Run the Parser px and pass it's result to the Parser fp.

(<|) ∷ (a → Parser s b) → Parser s a → Parser s b

Run the Parser px and pass it's result to the Parser fp.

(/>) ∷ Parser s a → Parser s b → Parser s b

Run the Parser p1 then the Parser p2.

(?>) ∷ Parser s a → Parser s a → Parser s a

Run the Parser p1 and, if it succeeds, return it's result, if it fails run the Parser p2.

convert ∷ a → Parser s a

Generate a Parser that when called will produce the value x.