Text.Cyp

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

parseCyp ∷ Parser a → Stream → Maybe a

Run the Parser p on the input Stream stream0.

parseCypPedantic ∷ Parser a → Stream → Maybe a

Run the Parser p on the input Stream stream0 and, additionaly, fail if their is any remaining input.

(>>!) ∷ Parser a → Parser a → Parser a

Run the Parser q if, and only if, the Parser p fails.

manyOf ∷ Parser a → Parser [a]

Match zero of more occurences of the Parser p, returning the concatenation of each result.

someOf ∷ Parser a → Parser [a]

Match one of more occurences of the Parser p, returning the concatenation of each result.