Text.Cyp.Bool

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchBool ∷ Parser Bool

Match a Bool.

matchLitBool ∷ Bool → Parser Bool

Match the Bool b.