Text.Cyp.Char

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

satisfy ∷ (Char → Bool) → Parser String Char

Match a Char that satisfies the satisfyicate p.

control ∷ Parser String Char

Match a control Char.

space ∷ Parser String Char

Match a whitespace Char.

lower ∷ Parser String Char

Match a lower case Char.

upper ∷ Parser String Char

Match an upper case Char.

alpha ∷ Parser String Char

Match an alphabetic Char.

alphaNum ∷ Parser String Char

Match an alphabetic Char or a digit Char.

printable ∷ Parser String Char

Match a printableable Char.

digit ∷ Parser String Char

Match an ASCII digit Char.

octDigit ∷ Parser String Char

Match an ASCII octal digit Char.

hexDigit ∷ Parser String Char

Match an ASCII hexadecimal digit Char.

letter ∷ Parser String Char

Match a letter Char.

mark ∷ Parser String Char

Match a mark Char.

number ∷ Parser String Char

Match a number Char.

punctuation ∷ Parser String Char

Match a punctuation symbol Char.

symbol ∷ Parser String Char

Match a symbol Char.

separator ∷ Parser String Char

Match a separator Char.

ascii ∷ Parser String Char

Match an ASCII Char

latin1 ∷ Parser String Char

Match a Latin-1 Char.

asciiUpper ∷ Parser String Char

Match an upper case ASCII Char.

asciiLower ∷ Parser String Char

Match a lower case ASCII Char.

char ∷ Parser String Char

Match any Char.

litChar ∷ Char → Parser String Char

Match the Char c.