Text.Cyp.Combinators

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

many ∷ Parser s a → Parser s [a]

Match zero or more occurences of the Parser p.

some ∷ Parser s a → Parser s [a]

Match one or more occurences of the Parser p.

between ∷ Parser s a → Parser s b → Parser s c → Parser s a

Match the Parser x between the Parsers x and y.

termBy ∷ Parser s a → Parser s b → Parser s a

Match one of the Parser p separated by the Parser q.

manySepBy ∷ Parser s a → Parser s b → Parser s [a]

Match zero or more of the Parser p separated by the Parser q.

someSepBy ∷ Parser s a → Parser s b → Parser s [a]

Match one or more of the Parser p separated by the Parser q.

manySepTermBy ∷ Parser s a → Parser s b → Parser s [a]

Match zero or more of the Parser p separated by the Parser q, followed by q.

someSepTermBy ∷ Parser s a → Parser s b → Parser s [a]

Match one or more of the Parser p separated by the Parser q, followed by q.

saveWith ∷ a → Parser s a → Parser s a

Attempt to match the Parser p, success resulting in p's result and failure resulting in a.