Text.Cyp.Float

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchFractional ∷ (Fractional f, Read f) ⇒ Parser f

matchFloat ∷ Parser Float

matchDouble ∷ Parser Double

matchLitFractional ∷ (Fractional f, Read f, Show f) ⇒ f → Parser f

matchLitFloat ∷ Float → Parser Float

matchLitDouble ∷ Double → Parser Double