Text.Cyp.Fractional

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

fractional ∷ (Fractional f, Read f) ⇒ Parser String f

Match a member of the Fractional class.

float ∷ Parser String Float

Match a Float.

double ∷ Parser String Double

Match a Double.

litFractional ∷ (Fractional f, Read f, Show f) ⇒ f → Parser String f

Match the member of the Fractional class, f.

litFloat ∷ Float → Parser String Float

Match the Float, f.

litDouble ∷ Double → Parser String Double

Match the Double, d.