Text.Cyp.Int

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchIntegral ∷ (Integral a, Read a) ⇒ Parser a

Match a member of the Integral class.

matchInt ∷ Parser Int

Match an Int.

matchInt8 ∷ Parser Int8

Match an Int8.

matchInt16 ∷ Parser Int16

Match an Int16.

matchInt32 ∷ Parser Int32

Match an Int32.

matchInt64 ∷ Parser Int64

Match an Int64.

matchLitIntegral ∷ (Integral a, Read a, Show a) ⇒ a → Parser a

Match the member of the Integral class, i.

matchLitInt ∷ Int → Parser Int

Match the Int i.

matchLitInt8 ∷ Int8 → Parser Int8

Match the Int8 i.

matchLitInt16 ∷ Int16 → Parser Int16

Match the Int16 i.

matchLitInt32 ∷ Int32 → Parser Int32

Match the Int32 i.

matchLitInt64 ∷ Int64 → Parser Int64

Match the Int64 i.