Text.Cyp.Integral

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

integral ∷ (Integral a, Read a) ⇒ Parser String a

Match a member of the Integral class.

int ∷ Parser String Int

Match an Int.

int8 ∷ Parser String Int8

Match an Int8.

int16 ∷ Parser String Int16

Match an Int16.

int32 ∷ Parser String Int32

Match an Int32.

int64 ∷ Parser String Int64

Match an Int64.

litIntegral ∷ (Integral a, Read a, Show a) ⇒ a → Parser String a

Match the member of the Integral class, i.

litInt ∷ Int → Parser String Int

Match the Int i.

litInt8 ∷ Int8 → Parser String Int8

Match the Int8 i.

litInt16 ∷ Int16 → Parser String Int16

Match the Int16 i.

litInt32 ∷ Int32 → Parser String Int32

Match the Int32 i.

litInt64 ∷ Int64 → Parser String Int64

Match the Int64 i.