Text.Cyp.Language

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

infixOp ∷ Parser String a → Parser String b → Parser String (b, a, b)

Match the Parser po between the Parser pa.

prefixOp ∷ Parser String a → Parser String b → Parser String (a, b)

Match the Parser po then the Parser pa.

identifier ∷ Parser String Char → Parser String Char → Parser String String

Match a identifier with the initial Char ed by pini and all of the remaining Chars ed by pres

parens ∷ Parser String a → Parser String a

Match the Parser p between parenthesis.

braces ∷ Parser String a → Parser String a

Match the Parser p between curly braces.

blocks ∷ Parser String a → Parser String a

Match the Parser p between block braces.

angles ∷ Parser String a → Parser String a

Match the Parser p between angle brackets.

doubleQuotes ∷ Parser String a → Parser String a

Match the Parser p between double quotes.

singleQuotes ∷ Parser String a → Parser String a

Match the Parser p between single quotes.

backQuotes ∷ Parser String a → Parser String a

Match the Parser p between back quotes.

commaSep ∷ Parser String a → Parser String [a]

Match many of the Parser p separated by commas.

commaSep1 ∷ Parser String a → Parser String [a]

Match some of the Parser p separated by commas.

semicolonSep ∷ Parser String a → Parser String [a]

Match many of the Parser p separated by commas.

semicolonSep1 ∷ Parser String a → Parser String [a]

Match some of the Parser p separated by commas.

doubleSmartQuotes ∷ Parser String a → Parser String a

Match the Parser p between smart double quotes.

singleSmartQuotes ∷ Parser String a → Parser String a

Match the Parser p between smart single quotes.