Text.Cyp.String

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

controls ∷ Parser String String

Match a String, of a length >= 0, of control Chars.

controls1 ∷ Parser String String

Match a String, of a length >= 1, of control Chars.

spaces ∷ Parser String String

Match a String, of a length >= 0, of white space Chars.

spaces1 ∷ Parser String String

Match a String, of a length >= 1, of white space Chars.

lowers ∷ Parser String String

Match a String, of a length >= 0, of lower case Chars.

lowers1 ∷ Parser String String

Match a String, of a length >= 1, of lower case Chars.

uppers ∷ Parser String String

Match a String, of a length >= 0, of upper case Chars.

uppers1 ∷ Parser String String

Match a String, of a length >= 1, of upper case Chars.

alphas ∷ Parser String String

Match a String, of a length >= 0, of alphabetic Chars.

alphas1 ∷ Parser String String

Match a String, of a length >= 1, of alphabetic Chars.

alphaNums ∷ Parser String String

Match a String, of a length >= 0, of alphabetic Chars and digit Chars.

alphaNums1 ∷ Parser String String

Match a String, of a length >= 1, of alphabetic Chars and digit Chars.

printables ∷ Parser String String

Match a String, of a length >= 0, of printableable Chars.

printables1 ∷ Parser String String

Match a String, of a length >= 1, of printableable Chars.

digits ∷ Parser String String

Match a String, of a length >= 0, of ASCII digit Chars.

digits1 ∷ Parser String String

Match a String, of a length >= 1, of ASCII digit Chars.

octDigits ∷ Parser String String

Match a String, of a length >= 0, of octal digit Chars.

octDigits1 ∷ Parser String String

Match a String, of a length >= 1, of octal digit Chars.

hexDigits ∷ Parser String String

Match a String, of a length >= 0, of ASCII hexadecimal digit Chars.

hexDigits1 ∷ Parser String String

Match a String, of a length >= 1, of ASCII hexadecimal digit Chars.

letters ∷ Parser String String

Match a String, of a length >= 0, of letter Chars.

letters1 ∷ Parser String String

Match a String, of a length >= 1, of letter Chars.

marks ∷ Parser String String

Match a String, of a length >= 0, of mark Chars.

marks1 ∷ Parser String String

Match a String, of a length >= 1, of mark Chars.

punctuations ∷ Parser String String

Match a String, of a length >= 0, of punctuation symbol Chars.

punctuations1 ∷ Parser String String

Match a String, of a length >= 1, of punctuation symbol Chars.

symbols ∷ Parser String String

Match a String, of a length >= 0, of symbol Chars.

symbols1 ∷ Parser String String

Match a String, of a length >= 1, of symbol Chars.

separators ∷ Parser String String

Match a String, of a length >= 0, of separator Chars.

separators1 ∷ Parser String String

Match a String, of a length >= 1, of separator Chars.

asciis ∷ Parser String String

Match a String, of a length >= 0, of ASCII Chars.

asciis1 ∷ Parser String String

Match a String, of a length >= 1, of ASCII Chars.

latin1s ∷ Parser String String

Match a String, of a length >= 0, of Latin-1 Chars.

latin1s1 ∷ Parser String String

Match a String, of a length >= 1, of Latin-1 Chars.

asciiUppers ∷ Parser String String

Match a String, of a length >= 0, of upper case ASCII Chars.

asciiUppers1 ∷ Parser String String

Match a String, of a length >= 1, of upper case ASCII Chars.

asciiLowers ∷ Parser String String

Match a String, of a length >= 0, of lower case ASCII Chars.

asciiLowers1 ∷ Parser String String

Match a String, of a length >= 1, of lower case ASCII Chars.

chars ∷ Parser String String

Match a String, of the length >= 0, of any Chars.

chars1 ∷ Parser String String

Match a String, of the length >= 1, of any Chars.

litString ∷ String → Parser String String

Match the String s.