Text.Cyp.Symbol

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchBang ∷ Parser String Char

Match a bang.

matchDoubleQuote ∷ Parser String Char

Match an double quote.

matchHash ∷ Parser String Char

Match a hash.

matchDollar ∷ Parser String Char

Match a dollar sign.

matchPercent ∷ Parser String Char

Match a percent sign.

matchConjunction ∷ Parser String Char

Match conjunction.

matchSingleQuote ∷ Parser String Char

Match single quote.

matchOpenParen ∷ Parser String Char

Match opening parenthesis.

matchCloseParen ∷ Parser String Char

Match closing parenthesis.

matchAsterisk ∷ Parser String Char

Match an asterisk.

matchPlus ∷ Parser String Char

Match plus sign.

matchComma ∷ Parser String Char

Match a comma.

matchMinus ∷ Parser String Char

Match a minus sign.

matchPeriod ∷ Parser String Char

Match a period.

matchForwardSlash ∷ Parser String Char

Match a forward slash.

matchColon ∷ Parser String Char

Match a colon.

matchSemicolon ∷ Parser String Char

Match semi-colon.

matchLessThan ∷ Parser String Char

Match a less than sign.

matchEquals ∷ Parser String Char

Match equals.

matchGreaterThan ∷ Parser String Char

Match an greater than sign.

matchQuestionMark ∷ Parser String Char

Match a question mark.

matchAt ∷ Parser String Char

Match an at sign.

matchOpenBlock ∷ Parser String Char

Match an open block.

matchBackwardSlash ∷ Parser String Char

Match a backslash.

matchCloseBlock ∷ Parser String Char

Match an close block.

matchHat ∷ Parser String Char

Match a hat.

matchUnderscore ∷ Parser String Char

Match an underscore.

matchBackQuote ∷ Parser String Char

Match a back quote.

matchOpenBrace ∷ Parser String Char

Match an opening curly brace.

matchCloseBrace ∷ Parser String Char

Match an closing curly brace.

matchPipe ∷ Parser String Char

Match a pipe.

matchTilda ∷ Parser String Char

Match a Tilda.

matchOpenSmartDoubleQuote ∷ Parser String Char

matchCloseSmartDoubleQuote ∷ Parser String Char

matchOpenSmartSingleQuote ∷ Parser String Char

matchCloseSmartSingleQuote ∷ Parser String Char