Data.Cyp

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

newtype Parser s a

The parser AGDT, which is a function that takes an initial input stream and returns, either, Just an a paired with a s or Nothing.

Constructors

P (s → Maybe (a, s))

empty ∷ Alternative f ⇒ ∀ a. f a

(<|>) ∷ Alternative f ⇒ ∀ a. f a → f a → f a

some ∷ Alternative f ⇒ ∀ a. f a → f [a]

many ∷ Alternative f ⇒ ∀ a. f a → f [a]

apply ∷ Parser s a → s → Maybe (a, s)

Remove a Parsers dummy constructor, evaluating to the Parser its self.