Text.Cyp

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

parse ∷ Parser s a → [s] → Maybe a

Run the Parser p on the input stream stream.

parsePedantic ∷ Parser s a → [s] → Maybe a

Run the Parser p on the input stream stream0 and fail if the resulting stream isn't empty.