Text.Cyp.Char

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchPred ∷ (Char → Bool) → Parser Char

Match a Char in the input Stream that satisfies, the predicate, p.

matchControl ∷ Parser Char

Match a control Char.

matchSpace ∷ Parser Char

Match a whitespace Char.

matchLower ∷ Parser Char

Match a lower case Char.

matchUpper ∷ Parser Char

Match an upper case Char.

matchAlpha ∷ Parser Char

Match an alphabetic Char.

matchAlphaNum ∷ Parser Char

Match an alphabetic Char or a digit Char.

matchPrint ∷ Parser Char

Match a printable Char.

matchDigit ∷ Parser Char

Match an ASCII digit Char.

matchOctDigit ∷ Parser Char

Match an ASCII octal digit Char.

matchHexDigit ∷ Parser Char

Match an ASCII hexadecimal digit Char.

matchLetter ∷ Parser Char

Match a letter Char.

matchMark ∷ Parser Char

Match a mark Char.

matchNumber ∷ Parser Char

Match a number Char.

matchPunctuation ∷ Parser Char

Match a punctuation symbol Char.

matchSymbol ∷ Parser Char

Match a symbol Char.

matchSeparator ∷ Parser Char

Match a separator Char.

matchAscii ∷ Parser Char

Match an ASCII Char

matchLatin1 ∷ Parser Char

Match a Latin-1 Char.

matchAsciiUpper ∷ Parser Char

Match an upper case ASCII Char.

matchAsciiLower ∷ Parser Char

Match a lower case ASCII Char.

matchAnyChar ∷ Parser Char

Match any Char.

matchChar ∷ Char → Parser Char

Match the Char c.