Text.Cyp.Combinators

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

switch ∷ [(Char → Bool, Parser String a)] → Parser String a

Take a list of pairs of conditionals and parser and execute the first parse who's paired conditioanl, when applied to the next Char in the strea, succeeds.

anything ∷ Char → Bool

Alwasy return true.

many ∷ Alternative f ⇒ ∀ a. f a → f [a]

some ∷ Alternative f ⇒ ∀ a. f a → f [a]

(<|>) ∷ Alternative f ⇒ ∀ a. f a → f a → f a

empty ∷ Alternative f ⇒ ∀ a. f a

between ∷ Parser s a → Parser s b → Parser s c → Parser s a

Match the Parser x between the Parsers x and y.

termBy ∷ Parser s a → Parser s b → Parser s a

Match one of the Parser p separated by the Parser q.

sepBys ∷ Parser s a → Parser s b → Parser s [a]

Match zero or more of the Parser p separated by the Parser q.

sepBys1 ∷ Parser s a → Parser s b → Parser s [a]

Match one or more of the Parser p separated by the Parser q.

sepTermBys ∷ Parser s a → Parser s b → Parser s [a]

Match zero or more of the Parser p separated by the Parser q, followed by q.

sepTermBys1 ∷ Parser s a → Parser s b → Parser s [a]

Match one or more of the Parser p separated by the Parser q, followed by q.

saveWith ∷ a → Parser s a → Parser s a

Attempt to match the Parser p, success resulting in p's result and failure resulting in a.