Text.Cyp.Float

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchFractional ∷ (Fractional f, Read f) ⇒ Parser f

Match a member of the Fractional class.

matchFloat ∷ Parser Float

Match a Float.

matchDouble ∷ Parser Double

Match a Double.

matchLitFractional ∷ (Fractional f, Read f, Show f) ⇒ f → Parser f

Match the member of the Fractional class, f.

matchLitFloat ∷ Float → Parser Float

Match the Float, f.

matchLitDouble ∷ Double → Parser Double

Match the Double, d.