Text.Cyp.Language

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchInfixOp ∷ Parser Char a → Parser Char b → Parser Char (b, a, b)

Match the Parser po between the Parser pa.

matchPrefixOp ∷ Parser Char a → Parser Char b → Parser Char (a, b)

Match the Parser po then the Parser pa.

matchIdentifier ∷ Parser Char Char → Parser Char Char → Parser Char String

Match a identifier with the initial Char matched by pini and all of the remaining Chars matched by pres

matchParens ∷ Parser Char a → Parser Char a

Match the Parser p between parenthesis.

matchBraces ∷ Parser Char a → Parser Char a

Match the Parser p between curly braces.

matchBlocks ∷ Parser Char a → Parser Char a

Match the Parser p between block braces.

matchAngles ∷ Parser Char a → Parser Char a

Match the Parser p between angle brackets.

matchDoubleQuotes ∷ Parser Char a → Parser Char a

Match the Parser p between double quotes.

matchSingleQuotes ∷ Parser Char a → Parser Char a

Match the Parser p between single quotes.

matchBackQuotes ∷ Parser Char a → Parser Char a

Match the Parser p between back quotes.

matchCommaSep ∷ Parser Char a → Parser Char [a]

Match many of the Parser p separated by commas.

matchCommaSep1 ∷ Parser Char a → Parser Char [a]

Match some of the Parser p separated by commas.

matchSemicolonSep ∷ Parser Char a → Parser Char [a]

Match many of the Parser p separated by commas.

matchSemicolonSep1 ∷ Parser Char a → Parser Char [a]

Match some of the Parser p separated by commas.