Text.Cyp.Language

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchInfixOp ∷ Parser Char a → Parser Char b → Parser Char (b, a, b)

matchPrefixOp ∷ Parser Char a → Parser Char b → Parser Char (a, b)

matchIdentifier ∷ Parser Char Char → Parser Char Char → Parser Char String

matchParens ∷ Parser Char a → Parser Char a

matchBraces ∷ Parser Char a → Parser Char a

matchBlocks ∷ Parser Char a → Parser Char a

matchAngles ∷ Parser Char a → Parser Char a

matchDoubleQuotes ∷ Parser Char a → Parser Char a

matchSingleQuotes ∷ Parser Char a → Parser Char a

matchBackQuotes ∷ Parser Char a → Parser Char a

matchCommaSep ∷ Parser Char a → Parser Char [a]

matchCommaSep1 ∷ Parser Char a → Parser Char [a]

matchSemicolonSep ∷ Parser Char a → Parser Char [a]

matchSemicolonSep1 ∷ Parser Char a → Parser Char [a]