Text.Cyp.String

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

matchControls ∷ Parser String

Match a String, of a length >= 0, of control Chars.

matchControls1 ∷ Parser String

Match a String, of a length >= 1, of control Chars.

matchSpaces ∷ Parser String

Match a String, of a length >= 0, of white space Chars.

matchSpaces1 ∷ Parser String

Match a String, of a length >= 1, of white space Chars.

matchLowers ∷ Parser String

Match a String, of a length >= 0, of lower case Chars.

matchLowers1 ∷ Parser String

Match a String, of a length >= 1, of lower case Chars.

matchUppers ∷ Parser String

Match a String, of a length >= 0, of upper case Chars.

matchUppers1 ∷ Parser String

Match a String, of a length >= 1, of upper case Chars.

matchAlphas ∷ Parser String

Match a String, of a length >= 0, of alphabetic Chars.

matchAlphas1 ∷ Parser String

Match a String, of a length >= 1, of alphabetic Chars.

matchAlphaNums ∷ Parser String

Match a String, of a length >= 0, of alphabetic Chars and digit Chars.

matchAlphaNums1 ∷ Parser String

Match a String, of a length >= 1, of alphabetic Chars and digit Chars.

matchPrints ∷ Parser String

Match a String, of a length >= 0, of printable Chars.

matchPrints1 ∷ Parser String

Match a String, of a length >= 1, of printable Chars.

matchDigits ∷ Parser String

Match a String, of a length >= 0, of ASCII digit Chars.

matchDigits1 ∷ Parser String

Match a String, of a length >= 1, of ASCII digit Chars.

matchOctDigits ∷ Parser String

Match a String, of a length >= 0, of octal digit Chars.

matchOctDigits1 ∷ Parser String

Match a String, of a length >= 1, of octal digit Chars.

matchHexDigits ∷ Parser String

Match a String, of a length >= 0, of ASCII hexadecimal digit Chars.

matchHexDigits1 ∷ Parser String

Match a String, of a length >= 1, of ASCII hexadecimal digit Chars.

matchLetters ∷ Parser String

Match a String, of a length >= 0, of letter Chars.

matchLetters1 ∷ Parser String

Match a String, of a length >= 1, of letter Chars.

matchMarks ∷ Parser String

Match a String, of a length >= 0, of mark Chars.

matchMarks1 ∷ Parser String

Match a String, of a length >= 1, of mark Chars.

matchPunctuations ∷ Parser String

Match a String, of a length >= 0, of punctuation symbol Chars.

matchPunctuations1 ∷ Parser String

Match a String, of a length >= 1, of punctuation symbol Chars.

matchSymbols ∷ Parser String

Match a String, of a length >= 0, of symbol Chars.

matchSymbols1 ∷ Parser String

Match a String, of a length >= 1, of symbol Chars.

matchSeparators ∷ Parser String

Match a String, of a length >= 0, of separator Chars.

matchSeparators1 ∷ Parser String

Match a String, of a length >= 1, of separator Chars.

matchAsciis ∷ Parser String

Match a String, of a length >= 0, of ASCII Chars.

matchAsciis1 ∷ Parser String

Match a String, of a length >= 1, of ASCII Chars.

matchLatin1s ∷ Parser String

Match a String, of a length >= 0, of Latin-1 Chars.

matchLatin1s1 ∷ Parser String

Match a String, of a length >= 1, of Latin-1 Chars.

matchAsciiUppers ∷ Parser String

Match a String, of a length >= 0, of upper case ASCII Chars.

matchAsciiUppers1 ∷ Parser String

Match a String, of a length >= 1, of upper case ASCII Chars.

matchAsciiLowers ∷ Parser String

Match a String, of a length >= 0, of lower case ASCII Chars.

matchAsciiLowers1 ∷ Parser String

Match a String, of a length >= 1, of lower case ASCII Chars.

matchString ∷ String → Parser String

Match the String s.