Text.Cyp.Symbol

Description

Copyright (c), Kove W. Ochre-Salter, 2018. This Haskell source-file is freely distributable under terms of the MIT License. A copy of said License is provided in the root directory of this project (LICENSE).

Documentation

bang ∷ Parser String Char

Match a bang.

doubleQuote ∷ Parser String Char

Match an double quote.

hash ∷ Parser String Char

Match a hash.

dollar ∷ Parser String Char

Match a dollar sign.

percent ∷ Parser String Char

Match a percent sign.

conjunction ∷ Parser String Char

Match conjunction.

singleQuote ∷ Parser String Char

Match single quote.

openParen ∷ Parser String Char

Match opening parenthesis.

closeParen ∷ Parser String Char

Match closing parenthesis.

asterisk ∷ Parser String Char

Match an asterisk.

plus ∷ Parser String Char

Match plus sign.

comma ∷ Parser String Char

Match a comma.

minus ∷ Parser String Char

Match a minus sign.

period ∷ Parser String Char

Match a period.

forwardSlash ∷ Parser String Char

Match a forward slash.

colon ∷ Parser String Char

Match a colon.

semicolon ∷ Parser String Char

Match semi-colon.

lessThan ∷ Parser String Char

Match a less than sign.

equals ∷ Parser String Char

Match equals.

greaterThan ∷ Parser String Char

Match an greater than sign.

questionMark ∷ Parser String Char

Match a question mark.

at ∷ Parser String Char

Match an at sign.

openBlock ∷ Parser String Char

Match an open block.

backwardSlash ∷ Parser String Char

Match a backslash.

closeBlock ∷ Parser String Char

Match an close block.

hat ∷ Parser String Char

Match a hat.

underscore ∷ Parser String Char

Match an underscore.

backQuote ∷ Parser String Char

Match a back quote.

openBrace ∷ Parser String Char

Match an opening curly brace.

closeBrace ∷ Parser String Char

Match an closing curly brace.

pipe ∷ Parser String Char

Match a pipe.

tilda ∷ Parser String Char

Match a Tilda.

openSmartDoubleQuote ∷ Parser String Char

closeSmartDoubleQuote ∷ Parser String Char

openSmartSingleQuote ∷ Parser String Char

closeSmartSingleQuote ∷ Parser String Char